refactor: Use timeserie instead of trip to get start/end place name
