+++ src/components/Trip/BottomSheet/BottomSheetContent.jsx
@@ -9,13 +9,11 @@ import { BottomSheetItem } from 'cozy-ui/transpiled/react/BottomSheet'
 
 import TimelineNode from 'src/components/Timeline/TimelineNode'
 import TimelineSections from 'src/components/Timeline/TimelineSections'
+import { getTripStartDate, getTripEndDate, formatDate } from 'src/lib/trips'
 import {
-  getStartPlaceDisplayName,
   getEndPlaceDisplayName,
-  getTripStartDate,
-  getTripEndDate,
-  formatDate
-} from 'src/lib/trips'
+  getStartPlaceDisplayName
+} from 'src/lib/timeseries'
 import { useTrip } from 'src/components/Providers/TripProvider'
 import PurposeItem from 'src/components/Trip/BottomSheet/PurposeItem'
 import PurposeEditDialog from 'src/components/EditDialogs/PurposeEditDialog'
@@ -28,12 +26,10 @@ const styles = {
 
 const BottomSheetContent = () => {
   const { f, lang } = useI18n()
-  const { trip } = useTrip()
+  const { trip, timeserie } = useTrip()
   const { isDesktop } = useBreakpoints()
   const [showPurposeDialog, setShowPurposeDialog] = useState(false)
 
-  const startPlaceName = useMemo(() => getStartPlaceDisplayName(trip), [trip])
-  const endPlaceName = useMemo(() => getEndPlaceDisplayName(trip), [trip])
   const startTime = useMemo(
     () => formatDate({ f, lang, date: getTripStartDate(trip) }),
     [f, lang, trip]
@@ -54,12 +50,16 @@ const BottomSheetContent = () => {
       <BottomSheetItem disableGutters>
         <Timeline className="u-pb-0 u-mb-0">
           <TimelineNode
-            label={startPlaceName}
+            label={getStartPlaceDisplayName(timeserie)}
             endLabel={startTime}
             type="start"
           />
           <TimelineSections />
-          <TimelineNode label={endPlaceName} endLabel={endTime} type="end" />
+          <TimelineNode
+            label={getEndPlaceDisplayName(timeserie)}
+            endLabel={endTime}
+            type="end"
+          />
         </Timeline>
       </BottomSheetItem>
       {/* TODO: Remove the Divider when we have the real desktop view */}
