+++ src/components/Trip/TripDialogMobile.jsx
@@ -1,18 +1,16 @@
-import React, { useMemo, useCallback, useRef } from 'react'
+import React, { useCallback, useRef } from 'react'
 import { useHistory } from 'react-router-dom'
 
 import { Dialog } from 'cozy-ui/transpiled/react/CozyDialogs'
 
-import { getEndPlaceDisplayName } from 'src/lib/trips'
+import { getEndPlaceDisplayName } from 'src/lib/timeseries'
 import { useTrip } from 'src/components/Providers/TripProvider'
 import TripDialogMobileContent from 'src/components/Trip/TripDialogMobileContent'
 
 const TripDialogMobile = () => {
-  const { trip } = useTrip()
+  const { timeserie } = useTrip()
   const history = useHistory()
   const titleRef = useRef(null)
-
-  const title = useMemo(() => getEndPlaceDisplayName(trip), [trip])
 
   const historyBack = useCallback(() => history.goBack(), [history])
 
@@ -22,7 +20,7 @@ const TripDialogMobile = () => {
       transitionDuration={0}
       disableGutters
       onBack={historyBack}
-      title={title}
+      title={getEndPlaceDisplayName(timeserie)}
       titleRef={titleRef}
       content={<TripDialogMobileContent titleRef={titleRef} />}
     />
